Cart API implemented

+++ database/migrations/2018_04_23_084654_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title')->unique();
-            $table->decimal('price', 4, 2);
+            $table->float('price');
             $table->timestamps();
         });
     }
